deleting team members

+++ src/pages/app/Team/Table/Index.vue
@@ -1,12 +1,12 @@
 <template>
   <md-table v-model="teamMembers" table-header-color="rose" class="table-striped table-hover">
     <md-table-row slot="md-table-row" slot-scope="{ item }">
-      <md-table-cell md-label="Name">{{ item.displayName }}</md-table-cell>
+      <md-table-cell md-label="Name">{{ item.firstName }} {{item.lastName}}</md-table-cell>
       <md-table-cell md-label="Email">{{ item.emailAddress }}</md-table-cell>
       <md-table-cell md-label="Role">{{ item.role }}</md-table-cell>
       <md-table-cell md-label="Permissions">{{ item.permissions }}</md-table-cell>
       <md-table-cell md-label="Actions">
-        <md-button class="md-raised md-primary md-icon-button" @click.native="deleteTeamMember(item)">
+        <md-button v-if="currentUserId !== item.id" class="md-raised md-primary md-icon-button" @click.native="deleteTeamMember(item)">
           <md-icon>delete</md-icon>
           <md-tooltip md-direction="top">Delete</md-tooltip>
         </md-button>
@@ -17,9 +17,11 @@
 
 <script>
   import {Modal, SimpleWizard, WizardTab} from "@/components";
+  import {mapState, mapGetters, mapMutations, mapActions} from 'vuex';
   import swal from "sweetalert2";
   import Teams from "@/models/Teams";
   import TeamMembers from "@/models/TeamMembers";
+  import indexVuexModule from "@/store/index";
 
   export default {
     components: {
@@ -46,7 +48,9 @@
     },
     data() {
       return {
-        inviteModalOpen: false
+        inviteModalOpen: false,
+        hideBtn: false,
+        currentUserId: indexVuexModule.state.currentUser.id,
       }
     },
     methods: {
@@ -67,21 +71,20 @@
       },
       async deleteTeam(id){
         let team = await Teams.first(id);
-        console.log('team');
         team.delete();
-        console.log(team, 'team');
         let teamIndex = this.teamMembers.findIndex(obj => obj.id === id)
 
         this.teamMembers.splice(teamIndex, 1)
 
       },
       async deleteTeamMember(teamMember){
-        console.log('teamMember');
-        teamMember.delete();
-        console.log(teamMember, 'teamMember');
-        let teamMemberIndex = this.teamMembers.findIndex(obj => obj.id === id)
+        if(indexVuexModule.state.currentUser.id !== teamMember.id){
+          teamMember.delete();
+          let teamMemberIndex = this.teamMembers.findIndex(obj => obj.id === teamMember.id)
 
-        this.teamMembers.splice(teamMemberIndex, 1)
+          this.teamMembers.splice(teamMemberIndex, 1)
+        }
+
 
       }
     }
